Move cache tests to own file
